requirements.txt
discord.py>=2.0.0
python-dotenv>=0.19.0
langchain>=0.0.267
langchain-google-genai>=0.0.1
google-generativeai>=0.3.0
Pillow>=9.0.0
aiohttp>=3.8.0
